+++ cantranslator/tests/bitfield_tests.cpp
@@ -74,7 +74,7 @@ START_TEST (test_get_off_byte_boundary)
 
 START_TEST (test_set_field)
 {
-    uint32_t data = 0;
+    uint64_t data = 0;
     setBitField(&data, 1, 0, 1);
     // unit32_t is stored in little endian but we read it in big endian, so the
     // retrieval in the set tests may look a little funky
@@ -95,7 +95,7 @@ END_TEST
 
 START_TEST (test_set_doesnt_clobber_existing_data)
 {
-    uint32_t data = 0xFFFC4DF3;
+    uint64_t data = 0xFFFC4DF3;
     setBitField(&data, 0x4fc8, 16, 16);
     // unit32_t is stored in little endian but we read it in big endian, so the
     // retrieval in the set tests may look a little funky
@@ -108,7 +108,7 @@ END_TEST
 
 START_TEST (test_set_off_byte_boundary)
 {
-    uint32_t data = 0xFFFC4DF3;
+    uint64_t data = 0xFFFC4DF3;
     setBitField(&data, 0x12, 12, 8);
     unsigned long result = getBitField((uint8_t*)&data, 8, 16);
     fail_unless(result == 0x2df1,
@@ -119,7 +119,7 @@ END_TEST
 
 START_TEST (test_set_odd_number_of_bits)
 {
-    uint32_t data = 0xFFFC4DF3;
+    uint64_t data = 0xFFFC4DF3;
     setBitField(&data, 0x12, 11, 5);
     unsigned long result = getBitField((uint8_t*)&data, 19, 5);
     fail_unless(result == 0x12,
